Title: 토끼와 경주, Time: 39ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240410/토끼와 경주/rabit-and-race.cpp
@@ -147,7 +147,7 @@ void change_dist(const int &pid, const bigint &L) {
  * O(P)
 */
 int best_rabbit() {
-    int bscore = 0;
+    bigint bscore = 0;
     for(int i = 0; i < P; i++) {
         if(score[i] + total - moved[i] > bscore) { bscore = score[i] + total - moved[i]; }
     }
